refactor: 메뉴그룹 분리
